Ouroboros.Network.ErrorPolicy module
Combine parts of `PeerState` and `Subscription.Ip` modules.
