+++ ouroboros-network/ouroboros-network.cabal
@@ -51,6 +51,7 @@ library
                        Ouroboros.Network.ChainFragment
                        Ouroboros.Network.Channel
                        Ouroboros.Network.DeltaQ
+                       Ouroboros.Network.ErrorPolicy
                        Ouroboros.Network.Mux
                        Ouroboros.Network.NodeToNode
                        Ouroboros.Network.NodeToClient
@@ -170,6 +171,7 @@ test-suite tests
                        Ouroboros.Network.MockChain.ProducerState
                        Ouroboros.Network.Channel
                        Ouroboros.Network.DeltaQ
+                       Ouroboros.Network.ErrorPolicy
                        Ouroboros.Network.Mux
                        Ouroboros.Network.MockNode
                        Ouroboros.Network.NodeToNode
