+++ ouroboros-network/src/Ouroboros/Network/Subscription/Worker.hs
@@ -54,6 +54,7 @@ import           Control.Monad.Class.MonadTimer
 import           Control.Monad.Class.MonadThrow
 import           Control.Tracer
 
+import           Ouroboros.Network.ErrorPolicy (CompleteApplication, Result (..))
 import           Ouroboros.Network.Server.ConnectionTable
 import           Ouroboros.Network.Subscription.Subscriber
 
@@ -77,36 +78,6 @@ maxConnectionAttemptDelay = 2 -- 2s delay
 --
 ipRetryDelay :: DiffTime
 ipRetryDelay = 10 -- 10s delay
-
-
--- | Result of the connection thread.  It's either result of an application, or
--- an exception thrown by it.
---
-data Result m addr r where
-     ApplicationResult
-       :: !(Time m)
-       -> !addr
-       -> !r
-       -> Result m addr r
-
-     Connected
-       :: !(Time m)
-       -> !addr
-       -> Result m addr r
-
-     ConnectionError
-       :: Exception e
-       => !(Time m)
-       -> !addr
-       -> !e
-       -> Result m addr r
-
-     ApplicationError
-       :: Exception e
-       => !(Time m)
-       -> !addr
-       -> !e
-       -> Result m addr r
 
 data ResOrAct m addr r =
      Res !(Result m addr r)
@@ -145,10 +116,6 @@ data SocketState m addr
 -- | Callback which fires: when we create or close a socket.
 --
 type SocketStateChange m s addr = SocketState m addr -> s -> STM m s
-
--- | Complete a connection, which receive application result (or exception).
---
-type CompleteApplication m s addr r = Result m addr r -> s -> STM m (s, m ())
 
 -- | Given current state 'retry' too keep the subscription worker going.
 -- When this transaction returns, all the threads spawned by the worker will be
